requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0flask-cors
pymongo
python-dotenv
openai
gunicorn==1.0.0
